Update modules, fix warnings

+++ src/c/include/global.h
@@ -95,8 +95,8 @@
 #define MASKED_READ(__value, __shift, __mask) (((__value) >> (__shift)) & (__mask))
 #define MASKED_WRITE(__to, __value, __shift, __mask) __to = (((__to) & ~((__mask) << (__shift))) | (((__value) & (__mask)) << (__shift)));
 
-#define PAGE_SIZE 0x1000
-#define ARC_STD_BUFF_SIZE 0x1000
+#define PAGE_SIZE (size_t)0x1000
+#define ARC_STD_BUFF_SIZE (size_t)0x1000
 
 extern struct ARC_BootMeta *Arc_BootMeta;
 extern struct ARC_TermMeta Arc_MainTerm;
